// src/auth/permissions/permission.enum.ts
enum Permission {
  Admin = 'Admin',
  Vendedor = 'Vendedor',
  LoginUpdate = 'LoginUpdate',

  UsuarioId = 'UsuarioId',
  UsuarioPost = 'UsuarioPost',
  UsuarioRead = 'UsuarioRead',
  UsuarioUpdate = 'UsuarioUpdate',
  UsuarioDelete = 'UsuarioDelete',

  ProdutoId = 'ProdutoId',
  ProdutoPost = 'ProdutoPost',
  ProdutoRead = 'ProdutoRead',
  ProdutoUpdate = 'ProdutoUpdate',
  ProdutoDelete = 'ProdutoDelete',
  ProdutoCategoria = 'ProdutoCategoria',

  ClienteId = 'ClienteId',
  ClientePost = 'ClientePost',
  ClienteRead = 'ClienteRead',
  ClienteUpdate = 'ClienteUpdate',
  ClienteDelete = 'ClienteDelete',
  ClienteMeus = 'ClienteMeus',
}

export default Permission;